// frontend/generated/models/ShippingInfoDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 8lind8box Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ShippingInfoDto
 */
export interface ShippingInfoDto {
    /**
     * 
     * @type {number}
     * @memberof ShippingInfoDto
     */
    shippingInfoId?: number;
    /**
     * 
     * @type {string}
     * @memberof ShippingInfoDto
     */
    address?: string;
    /**
     * 
     * @type {string}
     * @memberof ShippingInfoDto
     */
    ward?: string;
    /**
     * 
     * @type {string}
     * @memberof ShippingInfoDto
     */
    district?: string;
    /**
     * 
     * @type {string}
     * @memberof ShippingInfoDto
     */
    city?: string;
    /**
     * 
     * @type {string}
     * @memberof ShippingInfoDto
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof ShippingInfoDto
     */
    phoneNumber?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ShippingInfoDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof ShippingInfoDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ShippingInfoDto
     */
    updatedAt?: Date;
}

/**
 * Check if a given object implements the ShippingInfoDto interface.
 */
export function instanceOfShippingInfoDto(value: object): value is ShippingInfoDto {
    return true;
}

export function ShippingInfoDtoFromJSON(json: any): ShippingInfoDto {
    return ShippingInfoDtoFromJSONTyped(json, false);
}

export function ShippingInfoDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ShippingInfoDto {
    if (json == null) {
        return json;
    }
    return {
        
        'shippingInfoId': json['shippingInfoId'] == null ? undefined : json['shippingInfoId'],
        'address': json['address'] == null ? undefined : json['address'],
        'ward': json['ward'] == null ? undefined : json['ward'],
        'district': json['district'] == null ? undefined : json['district'],
        'city': json['city'] == null ? undefined : json['city'],
        'name': json['name'] == null ? undefined : json['name'],
        'phoneNumber': json['phoneNumber'] == null ? undefined : json['phoneNumber'],
        'isVisible': json['isVisible'] == null ? undefined : json['isVisible'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
    };
}

export function ShippingInfoDtoToJSON(json: any): ShippingInfoDto {
    return ShippingInfoDtoToJSONTyped(json, false);
}

export function ShippingInfoDtoToJSONTyped(value?: ShippingInfoDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'shippingInfoId': value['shippingInfoId'],
        'address': value['address'],
        'ward': value['ward'],
        'district': value['district'],
        'city': value['city'],
        'name': value['name'],
        'phoneNumber': value['phoneNumber'],
        'isVisible': value['isVisible'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
    };
}